Added new APIs about to test them all (course ones should all be working)

+++ MyCourseUTS.DataModel/ChoiceBlocks.cs
@@ -27,7 +27,7 @@ namespace MyCourseUTS.DataModel
         }
     
         public string ID { get; set; }
-        public int Version { get; set; }
+        public decimal Version { get; set; }
         public string Name { get; set; }
         public bool Active { get; set; }
         public int CreditPoints { get; set; }
